Move events contents to common
