+++ src/github.com/matrix-org/dendrite/clientapi/readers/profile.go
@@ -23,6 +23,7 @@ import (
 	"github.com/matrix-org/dendrite/clientapi/httputil"
 	"github.com/matrix-org/dendrite/clientapi/jsonerror"
 	"github.com/matrix-org/dendrite/clientapi/producers"
+	"github.com/matrix-org/dendrite/common"
 	"github.com/matrix-org/dendrite/common/config"
 	"github.com/matrix-org/dendrite/roomserver/api"
 	"github.com/matrix-org/gomatrixserverlib"
@@ -271,7 +272,7 @@ func buildMembershipEvents(
 			StateKey: &userID,
 		}
 
-		content := events.MemberContent{
+		content := common.MemberContent{
 			Membership: "join",
 		}
